Create and document the Model() object
Create and document a Model() object which will be the "type" of
objects scanned in from the JSON piece lists. The other properties of
this representation such as methods and effects lists is not something
which I have a concrete representation for as of yet but this is a
definite start.