Improvement: loading a malformatted JSON file now fails better
When attempting to parse a malforned JSON data file, Cortex will now
plainly state the failure and list the file, line and column of the
parse error. I expect that it will be outside of most users' ability
to debug such a failure but at least the failure is not a scary
looking traceback into a crash.